requirements.txt
ai2thor==2.1.0
alfworld==0.3.3
annotated-types==0.6.0
anyio==4.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
distro==1.9.0
exceptiongroup==1.2.0
fast-downward-textworld==20.6.1
filelock==3.13.3
Flask==2.1.3
fsspec==2024.3.1
h11==0.14.0
hashids==1.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
jericho==3.1.2
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
mementos==1.3.1
more-itertools==10.2.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.14.3
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
preshed==3.0.9
progressbar2==4.4.2
prompt-toolkit==3.0.43
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
ratelimit==2.2.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
TatSu==5.8.3
termcolor==2.4.0
textworld==1.6.1
thinc==8.2.3
tiktoken==0.6.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==2.0.3
zipp==3.18.1
